# livros-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Livros",
        "description": "Documentação da API de Livros",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor Local"
        }
    ],
    "paths": {
        "/api/assuntos": {
            "get": {
                "tags": [
                    "Assuntos"
                ],
                "summary": "Lista todos os assuntos",
                "operationId": "6494dc2c13d5ebcdbf4222925db1a646",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da página",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "descricao",
                        "in": "query",
                        "description": "Filtrar pela descrição do assunto",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de assuntos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Assunto"
                                            }
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Assuntos"
                ],
                "summary": "Cadastra um novo assunto",
                "operationId": "695b52026e2a76bf59f8450c7cc05479",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Assunto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Assunto cadastrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Assunto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/assuntos/{id}": {
            "get": {
                "tags": [
                    "Assuntos"
                ],
                "summary": "Exibe um assunto",
                "operationId": "14b955e5d304ecefb8633b95183c67f2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Assunto encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Assunto"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Assuntos"
                ],
                "summary": "Atualiza um assunto",
                "operationId": "adc52c6c83e59febfb4f9c965755d7e8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Assunto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Assunto atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Assunto"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Assuntos"
                ],
                "summary": "Remove um assunto",
                "operationId": "066f8c0ddcb8dc966758fdcb63ff7ca2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Assunto removido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/autors": {
            "get": {
                "tags": [
                    "Autores"
                ],
                "summary": "Lista todos os autores",
                "operationId": "9f30e4c6a9c77aa23256cd2d987840de",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da página",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "nome",
                        "in": "query",
                        "description": "Filtrar pelo nome do autor",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de autores",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Autor"
                                            }
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Autores"
                ],
                "summary": "Cadastra um novo autor",
                "operationId": "6bc8c9bf141c823e9c9b8eb193d0f9a0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Autor"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Autor cadastrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Autor"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/autors/{id}": {
            "get": {
                "tags": [
                    "Autores"
                ],
                "summary": "Exibe um autor",
                "operationId": "5dfc968217a48b09d254e61f80277b51",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Autor encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Autor"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Autores"
                ],
                "summary": "Atualiza um autor",
                "operationId": "942d8dce0dff12d1e21dcef3da708954",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Autor"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Autor atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Autor"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Autores"
                ],
                "summary": "Remove um autor",
                "operationId": "b945c300ee67d02fc55108bcd57a913c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Autor removido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Assunto": {
                "title": "Assunto",
                "description": "Modelo de Assunto",
                "required": [
                    "descricao"
                ],
                "properties": {
                    "cod_assunto": {
                        "description": "Código do assunto",
                        "type": "integer",
                        "example": 1
                    },
                    "descricao": {
                        "description": "Descrição do assunto",
                        "type": "string",
                        "example": "Literatura Brasileira"
                    },
                    "livros": {
                        "description": "Lista de livros relacionados ao assunto",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "type": "object"
            },
            "Autor": {
                "title": "Autor",
                "description": "Modelo de Autor",
                "required": [
                    "nome"
                ],
                "properties": {
                    "cod_autor": {
                        "description": "Código do autor",
                        "type": "integer",
                        "example": 1
                    },
                    "nome": {
                        "description": "Nome do autor",
                        "type": "string",
                        "example": "Machado de Assis"
                    },
                    "livros": {
                        "description": "Lista de livros do autor",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}